Popup warning too verbose
Removed popup warning when we try to format a file that is not a style file (css, etc). console.error(...) should be sufficient.
This is important because if you keybind this extensions format command you will get the warning popup when you format other files which is not what you want, you just want to format whatever file you're in with another formatter as expected.